+++ src/Web/Internal/HttpApiData.hs
@@ -13,67 +13,70 @@
 -- such as URL pieces, headers and query parameters.
 module Web.Internal.HttpApiData where
 
-#if __GLASGOW_HASKELL__ < 710
-import           Control.Applicative
-import           Data.Foldable                (Foldable)
-import           Data.Traversable             (Traversable (traverse))
-#endif
+import           Prelude                      ()
+import           Prelude.Compat
 
 import           Control.Arrow                (left, (&&&))
 import           Control.Monad                ((<=<))
-
+import qualified Data.Attoparsec.Text         as Atto
+import qualified Data.Attoparsec.Time         as Atto
 import           Data.ByteString              (ByteString)
 import qualified Data.ByteString              as BS
+import qualified Data.ByteString.Builder      as BS
 import qualified Data.ByteString.Lazy         as LBS
-import           Data.Monoid
-
+import           Data.Coerce                  (coerce)
+import           Data.Data                    (Data)
 import qualified Data.Fixed                   as F
-import           Data.Int
-import           Data.Word
-
+import           Data.Int                     (Int16, Int32, Int64, Int8)
+import           Data.Monoid                  (All (..), Any (..), Dual (..),
+                                               First (..), Last (..),
+                                               Product (..), Sum (..))
+import           Data.Semigroup               (Semigroup (..))
+import qualified Data.Semigroup               as Semi
+import           Data.Tagged                  (Tagged (..))
 import           Data.Text                    (Text)
 import qualified Data.Text                    as T
-import           Data.Text.Encoding           (decodeUtf8With, decodeUtf8', encodeUtf8)
+import           Data.Text.Encoding           (decodeUtf8', decodeUtf8With,
+                                               encodeUtf8)
 import           Data.Text.Encoding.Error     (lenientDecode)
 import qualified Data.Text.Lazy               as L
 import           Data.Text.Read               (Reader, decimal, rational,
                                                signed)
-
-import           Data.Time
-#if __GLASGOW_HASKELL__ < 710
-import           Data.Time.Locale.Compat
-#endif
-import           Data.Version
-
-#if MIN_VERSION_base(4,8,0)
-import           Data.Void
-import           Numeric.Natural
-#endif
-
+import           Data.Time                    (Day, FormatTime, LocalTime,
+                                               NominalDiffTime, TimeOfDay,
+                                               UTCTime, ZonedTime, formatTime)
+import           Data.Time.Locale.Compat      (defaultTimeLocale,
+                                               iso8601DateFormat)
+import           Data.Typeable                (Typeable)
+import qualified Data.UUID.Types              as UUID
+import           Data.Version                 (Version, parseVersion,
+                                               showVersion)
+import           Data.Void                    (Void, absurd)
+import           Data.Word                    (Word16, Word32, Word64, Word8)
+import qualified Network.HTTP.Types           as H
+import           Numeric.Natural              (Natural)
 import           Text.ParserCombinators.ReadP (readP_to_S)
 import           Text.Read                    (readMaybe)
+import           Web.Cookie                   (SetCookie, parseSetCookie,
+                                               renderSetCookie)
+
+#if MIN_VERSION_time(1,9,1)
+import           Data.Time                    (nominalDiffTimeToSeconds,
+                                               secondsToNominalDiffTime)
+#endif
 
 #if USE_TEXT_SHOW
 import           TextShow                     (TextShow, showt)
 #endif
 
-import qualified Data.UUID.Types              as UUID
 
-import qualified Data.ByteString.Builder      as BS
-import           Data.Data                    (Data)
-import           Data.Typeable                (Typeable)
-import qualified Network.HTTP.Types           as H
-
-import qualified Data.Attoparsec.Text         as Atto
-import qualified Data.Attoparsec.Time         as Atto
-
-import           Web.Cookie                   (SetCookie, parseSetCookie,
-                                               renderSetCookie)
 
 
 -- $setup
 -- >>> data BasicAuthToken = BasicAuthToken Text deriving (Show)
 -- >>> instance FromHttpApiData BasicAuthToken where parseHeader h = BasicAuthToken <$> parseHeaderWithPrefix "Basic " h; parseQueryParam p = BasicAuthToken <$> parseQueryParam p
+-- >>> import Data.Time
+-- >>> import Data.Version
 
 -- | Convert value to HTTP API data.
 --
@@ -432,10 +435,8 @@ instance ToHttpApiData Version where
   toUrlPiece = T.pack . showVersion
   toEncodedUrlPiece = unsafeToEncodedUrlPiece
 
-#if MIN_VERSION_base(4,8,0)
 instance ToHttpApiData Void    where toUrlPiece = absurd
 instance ToHttpApiData Natural where toUrlPiece = showt; toEncodedUrlPiece = unsafeToEncodedUrlPiece
-#endif
 
 instance ToHttpApiData Bool     where toUrlPiece = showTextData; toEncodedUrlPiece = unsafeToEncodedUrlPiece
 instance ToHttpApiData Ordering where toUrlPiece = showTextData; toEncodedUrlPiece = unsafeToEncodedUrlPiece
@@ -494,36 +495,77 @@ instance ToHttpApiData UTCTime where
   toUrlPiece = timeToUrlPiece "%H:%M:%S%QZ"
   toEncodedUrlPiece = unsafeToEncodedUrlPiece
 
+-- The CPP in both this function and the function after it are to avoid
+-- exporting @nominalDiffTimeToSeconds@ and @secondsToNominalDiffTime@,
+-- since these names are already used by @Data.Time@ from the @time@ library
+-- starting in version @1.9.1@.
+nominalDiffTimeToSecs :: NominalDiffTime -> F.Pico
+nominalDiffTimeToSecs =
+#if !MIN_VERSION_time(1,9,1)
+    realToFrac
+#else
+    nominalDiffTimeToSeconds
+#endif
+
+secsToNominalDiffTime :: F.Pico -> NominalDiffTime
+secsToNominalDiffTime =
+#if !MIN_VERSION_time(1,9,1)
+    realToFrac
+#else
+    secondsToNominalDiffTime
+#endif
+
 instance ToHttpApiData NominalDiffTime where
-  toUrlPiece = toUrlPiece . (floor :: NominalDiffTime -> Integer)
+  toUrlPiece = toUrlPiece . nominalDiffTimeToSecs
   toEncodedUrlPiece = unsafeToEncodedUrlPiece
 
 instance ToHttpApiData String   where toUrlPiece = T.pack
 instance ToHttpApiData Text     where toUrlPiece = id
 instance ToHttpApiData L.Text   where toUrlPiece = L.toStrict
 
-instance ToHttpApiData All where toUrlPiece = toUrlPiece . getAll; toEncodedUrlPiece = toEncodedUrlPiece . getAll
-instance ToHttpApiData Any where toUrlPiece = toUrlPiece . getAny; toEncodedUrlPiece = toEncodedUrlPiece . getAny
+instance ToHttpApiData All where
+  toUrlPiece        = coerce (toUrlPiece :: Bool -> Text)
+  toEncodedUrlPiece = coerce (toEncodedUrlPiece :: Bool -> BS.Builder)
+
+instance ToHttpApiData Any where
+  toUrlPiece        = coerce (toUrlPiece :: Bool -> Text)
+  toEncodedUrlPiece = coerce (toEncodedUrlPiece :: Bool -> BS.Builder)
 
 instance ToHttpApiData a => ToHttpApiData (Dual a) where
-  toUrlPiece = toUrlPiece . getDual
-  toEncodedUrlPiece = toEncodedUrlPiece . getDual
+  toUrlPiece        = coerce (toUrlPiece :: a -> Text)
+  toEncodedUrlPiece = coerce (toEncodedUrlPiece :: a -> BS.Builder)
 
 instance ToHttpApiData a => ToHttpApiData (Sum a) where
-  toUrlPiece = toUrlPiece . getSum
-  toEncodedUrlPiece = toEncodedUrlPiece . getSum
+  toUrlPiece        = coerce (toUrlPiece :: a -> Text)
+  toEncodedUrlPiece = coerce (toEncodedUrlPiece :: a -> BS.Builder)
 
 instance ToHttpApiData a => ToHttpApiData (Product a) where
-  toUrlPiece = toUrlPiece . getProduct
-  toEncodedUrlPiece = toEncodedUrlPiece . getProduct
+  toUrlPiece        = coerce (toUrlPiece :: a -> Text)
+  toEncodedUrlPiece = coerce (toEncodedUrlPiece :: a -> BS.Builder)
 
 instance ToHttpApiData a => ToHttpApiData (First a) where
-  toUrlPiece = toUrlPiece . getFirst
-  toEncodedUrlPiece = toEncodedUrlPiece . getFirst
+  toUrlPiece        = coerce (toUrlPiece :: Maybe a -> Text)
+  toEncodedUrlPiece = coerce (toEncodedUrlPiece :: Maybe a -> BS.Builder)
 
 instance ToHttpApiData a => ToHttpApiData (Last a) where
-  toUrlPiece = toUrlPiece . getLast
-  toEncodedUrlPiece = toEncodedUrlPiece . getLast
+  toUrlPiece        = coerce (toUrlPiece :: Maybe a -> Text)
+  toEncodedUrlPiece = coerce (toEncodedUrlPiece :: Maybe a -> BS.Builder)
+
+instance ToHttpApiData a => ToHttpApiData (Semi.Min a) where
+  toUrlPiece        = coerce (toUrlPiece :: a -> Text)
+  toEncodedUrlPiece = coerce (toEncodedUrlPiece :: a -> BS.Builder)
+
+instance ToHttpApiData a => ToHttpApiData (Semi.Max a) where
+  toUrlPiece        = coerce (toUrlPiece :: a -> Text)
+  toEncodedUrlPiece = coerce (toEncodedUrlPiece :: a -> BS.Builder)
+
+instance ToHttpApiData a => ToHttpApiData (Semi.First a) where
+  toUrlPiece        = coerce (toUrlPiece :: a -> Text)
+  toEncodedUrlPiece = coerce (toEncodedUrlPiece :: a -> BS.Builder)
+
+instance ToHttpApiData a => ToHttpApiData (Semi.Last a) where
+  toUrlPiece        = coerce (toUrlPiece :: a -> Text)
+  toEncodedUrlPiece = coerce (toEncodedUrlPiece :: a -> BS.Builder)
 
 -- |
 -- >>> toUrlPiece (Just "Hello")
@@ -555,6 +597,12 @@ instance ToHttpApiData SetCookie where
   toHeader = LBS.toStrict . BS.toLazyByteString . renderSetCookie
   -- toEncodedUrlPiece = renderSetCookie -- doesn't do things.
 
+instance ToHttpApiData a => ToHttpApiData (Tagged b a) where
+  toUrlPiece        = coerce (toUrlPiece :: a -> Text)
+  toHeader          = coerce (toHeader :: a -> ByteString)
+  toQueryParam      = coerce (toQueryParam :: a -> Text)
+  toEncodedUrlPiece = coerce (toEncodedUrlPiece ::  a -> BS.Builder)
+
 -- |
 -- >>> parseUrlPiece "_" :: Either Text ()
 -- Right ()
@@ -566,7 +614,7 @@ instance FromHttpApiData Char where
   parseUrlPiece s =
     case T.uncons s of
       Just (c, s') | T.null s' -> pure c
-      _                        -> defaultParseError s
+      _            -> defaultParseError s
 
 -- |
 -- >>> showVersion <$> parseUrlPiece "1.2.3"
@@ -577,7 +625,6 @@ instance FromHttpApiData Version where
       ((x, ""):_) -> pure x
       _           -> defaultParseError s
 
-#if MIN_VERSION_base(4,8,0)
 -- | Parsing a @'Void'@ value is always an error, considering @'Void'@ as a data type with no constructors.
 instance FromHttpApiData Void where
   parseUrlPiece _ = Left "Void cannot be parsed!"
@@ -588,7 +635,6 @@ instance FromHttpApiData Natural where
     if n < 0
       then Left ("underflow: " <> s <> " (should be a non-negative integer)")
       else Right (fromInteger n)
-#endif
 
 instance FromHttpApiData Bool     where parseUrlPiece = parseBoundedUrlPiece
 instance FromHttpApiData Ordering where parseUrlPiece = parseBoundedUrlPiece
@@ -640,16 +686,21 @@ instance FromHttpApiData ZonedTime where parseUrlPiece = runAtto Atto.zonedTime
 -- Right 2015-10-03 00:14:24 UTC
 instance FromHttpApiData UTCTime   where parseUrlPiece = runAtto Atto.utcTime
 
-instance FromHttpApiData NominalDiffTime where parseUrlPiece = fmap fromInteger . parseUrlPiece
+instance FromHttpApiData NominalDiffTime where parseUrlPiece = fmap secsToNominalDiffTime . parseUrlPiece
 
-instance FromHttpApiData All where parseUrlPiece = fmap All . parseUrlPiece
-instance FromHttpApiData Any where parseUrlPiece = fmap Any . parseUrlPiece
+instance FromHttpApiData All where parseUrlPiece = coerce (parseUrlPiece :: Text -> Either Text Bool)
+instance FromHttpApiData Any where parseUrlPiece = coerce (parseUrlPiece :: Text -> Either Text Bool)
 
-instance FromHttpApiData a => FromHttpApiData (Dual a)    where parseUrlPiece = fmap Dual    . parseUrlPiece
-instance FromHttpApiData a => FromHttpApiData (Sum a)     where parseUrlPiece = fmap Sum     . parseUrlPiece
-instance FromHttpApiData a => FromHttpApiData (Product a) where parseUrlPiece = fmap Product . parseUrlPiece
-instance FromHttpApiData a => FromHttpApiData (First a)   where parseUrlPiece = fmap First   . parseUrlPiece
-instance FromHttpApiData a => FromHttpApiData (Last a)    where parseUrlPiece = fmap Last    . parseUrlPiece
+instance FromHttpApiData a => FromHttpApiData (Dual a)    where parseUrlPiece = coerce (parseUrlPiece :: Text -> Either Text a)
+instance FromHttpApiData a => FromHttpApiData (Sum a)     where parseUrlPiece = coerce (parseUrlPiece :: Text -> Either Text a)
+instance FromHttpApiData a => FromHttpApiData (Product a) where parseUrlPiece = coerce (parseUrlPiece :: Text -> Either Text a)
+instance FromHttpApiData a => FromHttpApiData (First a)   where parseUrlPiece = coerce (parseUrlPiece :: Text -> Either Text (Maybe a))
+instance FromHttpApiData a => FromHttpApiData (Last a)    where parseUrlPiece = coerce (parseUrlPiece :: Text -> Either Text (Maybe a))
+
+instance FromHttpApiData a => FromHttpApiData (Semi.Min a)    where parseUrlPiece = coerce (parseUrlPiece :: Text -> Either Text a)
+instance FromHttpApiData a => FromHttpApiData (Semi.Max a)    where parseUrlPiece = coerce (parseUrlPiece :: Text -> Either Text a)
+instance FromHttpApiData a => FromHttpApiData (Semi.First a)  where parseUrlPiece = coerce (parseUrlPiece :: Text -> Either Text a)
+instance FromHttpApiData a => FromHttpApiData (Semi.Last a)   where parseUrlPiece = coerce (parseUrlPiece :: Text -> Either Text a)
 
 -- |
 -- >>> parseUrlPiece "Just 123" :: Either Text (Maybe Int)
@@ -700,6 +751,11 @@ instance FromHttpApiData SetCookie where
   parseUrlPiece = parseHeader  . encodeUtf8
   parseHeader   = Right . parseSetCookie
 
+instance FromHttpApiData a => FromHttpApiData (Tagged b a) where
+  parseUrlPiece   = coerce (parseUrlPiece :: Text -> Either Text a)
+  parseHeader     = coerce (parseHeader :: ByteString -> Either Text a)
+  parseQueryParam = coerce (parseQueryParam :: Text -> Either Text a)
+
 -------------------------------------------------------------------------------
 -- Attoparsec helpers
 -------------------------------------------------------------------------------
@@ -708,3 +764,5 @@ runAtto :: Atto.Parser a -> Text -> Either Text a
 runAtto p t = case Atto.parseOnly (p <* Atto.endOfInput) t of
     Left err -> Left (T.pack err)
     Right x  -> Right x
+
+